refactor(types): action.ts 完全削除 (#1186 Phase 3 完遂)
ISSUE #1186 Phase 3: 全 consumer (99 file) を types/v3 + utils/processFlowMetadata に
完全移行完了。本 PR で `frontend/src/types/action.ts` を物理削除。

## 移行 10 test file

action.*.test.ts の 10 file は `from "./action"` で隣接相対 import していたため
sed で `from "../types/v3"` に置換 (Phase 2-E の bulk sed は \"types/action\" pattern のみ)。
3 file は STEP_TYPE_LABELS / EXTERNAL_CALL_OUTCOME_VALUES の import path を
processFlowMetadata に補正。

## 削除内容

- frontend/src/types/action.ts (196 行、@deprecated 互換層) を削除
- ファイル名 action.*.test.ts は backward compat 維持 (内部 import は v3 直接)

## #1186 全 Phase 完遂サマリー

| Phase | PR | 内容 |
|---|---|---|
| 1 | #1192 | dead enum 削除 + JSDoc 同期 |
| 2-A | #1199 | UI metadata 抽出 + PrimaryKeyConstraint TS sync |
| 2-B | #1200 | 14 consumer 移行 |
| 2-C | #1201 | 6 consumer + 副次 silent bug fix |
| 2-D | #1202 | 46 file 一括移行 + v3 legacy alias 拡張 |
| 2-E/F/G | #1203 | 40 file 一括移行 + 9 ts-nocheck 追加 |
| 3 | 本 PR | action.ts 削除 + 10 test file 移行 |

最終状態:
- action.ts → 削除
- types/v3 → 全 legacy alias を re-export
- utils/processFlowMetadata → UI 表示メタデータ集約

## Test plan

- [x] frontend build (tsc + vite) → OK
- [x] frontend test → 2346 / 2346 pass
- [x] backend test → 516 / 516 pass
- [x] validate:samples (retail) → 7/7 flows pass

## 後続課題 (別 ISSUE 候補)

ts-nocheck 残存 file の proper type narrow は #1016 で継続 (本 ISSUE 範囲外)。
v3 legacy alias 群 (ProcessFlowType / StepType / SecretRef 等) は将来 v4 schema 移行時に
削除候補だが、現状は frontend consumer の名称互換維持のため温存。

Closes #1186
